requirements.txt
streamlit
pandas
plotly
google-cloud-bigquery
folium
python-dotenv
matplotlib